updated SingleMotor.lua to have assert
Its a simple change but overall helps in case you forget to format the options table which I have done on a few occasions. It was difficult to figure out what the issue was since the  traceback leads to the signal class so I thought a couple asserts in the setGoal function would help debugging in the future.